+++ src/tests/all_of.cpp
@@ -1,13 +1,13 @@
-// ============================== NONE OF TESTS =============================== //
+// ============================== ALL OF TESTS =============================== //
 // Project:         The Experimental Bit Algorithms Library
-// Name:            none_of.hpp
-// Description:     Tests for none_of algorithm for bit iterators
+// Name:            all_of.hpp
+// Description:     Tests for all_of algorithm for bit iterators
 // Creator:         Vincent Reverdy
 // Contributor(s):  Collin Gress [2019]
 // License: BSD 3-Clause License
 // ========================================================================== //
-#ifndef _NONE_OF_TESTS_HPP_INCLUDED
-#define _NONE_OF_TESTS_HPP_INCLUDED
+#ifndef _ALL_OF_TESTS_HPP_INCLUDED
+#define _ALL_OF_TESTS_HPP_INCLUDED
 // ========================================================================== //
 
 
@@ -15,23 +15,24 @@
 // =============================== PREAMBLE ================================= //
 // C++ standard library
 // Project sources
-#include "test_root.cc"
+#include "catch2.hpp"
 #include "bit.hpp"
+#include "test_utils.hpp"
 // Third-party libraries
 // Miscellaneous
 // ========================================================================== //
 
-TEMPLATE_TEST_CASE("none_of: is correct for trivial single word cases", "[none_of]", 
+TEMPLATE_TEST_CASE("all_of: is correct for trivial single word cases", "[all_of]", 
     unsigned short, unsigned int, unsigned long, unsigned long long) {
 
     TestType t = 0;
     bit::bit_iterator<TestType*> first(&t, 0);
     bit::bit_iterator<TestType*> last(&t + 1, 0);
 
-    auto predicate = [](bit::bit_value bv) { return bv == bit::bit1; };
+    auto predicate = [](bit::bit_value bv) { return bv == bit::bit0; };
 
-    REQUIRE(bit::none_of(first, last, predicate) 
-        == std::none_of(first, last, predicate));
+    REQUIRE(bit::all_of(first, last, predicate) 
+        == std::all_of(first, last, predicate));
 }
 
 #endif
