+++ src/tests/bit_details.cpp
@@ -16,6 +16,8 @@
 // ============================== PREAMBLE ================================== //
 // C++ standard library
 // Project sources
+#include "bit.hpp"
+#include "test_utils.hpp"
 // Third-party libraries
 #include "catch2.hpp"
 // ========================================================================== //
