+++ src/tests/input_container.cpp
@@ -15,8 +15,9 @@
 // C++ standard library
 #include <deque>
 // Project sources
-#include "test_root.cc"
+#include "catch2.hpp"
 #include "bit.hpp"
+#include "test_utils.hpp"
 // Third-party libraries
 // Miscellaneous
 // ========================================================================== //
