+++ src/tests/min_element.cpp
@@ -16,8 +16,9 @@
 // C++ standard library
 #include <iostream>
 // Project sources
-#include "test_root.cc"
+#include "catch2.hpp"
 #include "bit.hpp"
+#include "test_utils.hpp"
 // Third-party libraries
 // Miscellaneous
 // ========================================================================== //
