+++ src/tests/mismatch.cpp
@@ -16,8 +16,9 @@
 // =============================== PREAMBLE ================================= //
 // C++ standard library
 // Project sources
-#include "test_root.cc"
+#include "catch2.hpp"
 #include "bit.hpp"
+#include "test_utils.hpp"
 // Third-party libraries
 // Miscellaneous
 // ========================================================================== //
