+++ src/tests/replace.cpp
@@ -1,13 +1,13 @@
-// ============================= REVERSE TESTS =============================== //
+// ============================ REPLACE TESTS =============================== //
 // Project:         The Experimental Bit Algorithms Library
-// Name:            reverse.hpp
-// Description:     Tests for reverse algorithms 
+// Name:            replace.hpp
+// Description:     Tests for replace
 // Creator:         Vincent Reverdy
 // Contributor(s):  Bryce Kille [2019]
 // License:         BSD 3-Clause License
 // ========================================================================== //
-#ifndef _REVERSE_TESTS_HPP_INCLUDED
-#define _REVERSE_TESTS_HPP_INCLUDED
+#ifndef _REPLACE_TESTS_HPP_INCLUDED
+#define _REPLACE_TESTS_HPP_INCLUDED
 // ========================================================================== //
 
 
@@ -15,18 +15,19 @@
 // ============================== PREAMBLE ================================== //
 // C++ standard library
 // Project sources
-#include "test_root.cc"
+#include "catch2.hpp"
 #include "bit.hpp"
+#include "test_utils.hpp"
 // Third-party libraries
 // Miscellaneous
 // ========================================================================== //
 
 
 
-// ----------------------------- Reverse Tests ------------------------------- //
-TEMPLATE_PRODUCT_TEST_CASE("Reverse: single_word", 
+// ---------------------------- REPLACE Tests ----------------------------- //
+TEMPLATE_PRODUCT_TEST_CASE("Replace: single_word", 
                            "[template][product]", 
-                           (std::vector, std::list), 
+                           (std::vector, std::list, std::forward_list), 
                            (unsigned char, unsigned short, 
                             unsigned int, unsigned long)) {
 
@@ -42,56 +43,58 @@ TEMPLATE_PRODUCT_TEST_CASE("Reverse: single_word",
     auto bool_first = std::begin(boolcont);
     auto bool_last = std::end(boolcont);
 
-    bit::reverse(
+    bit::replace(
         bfirst, 
-        std::next(bfirst, 3)
+        std::next(bfirst, 3), 
+        bit::bit0,
+        bit::bit1
     );
-    std::reverse(
+    std::replace(
         bool_first, 
-        std::next(bool_first, 3)
+        std::next(bool_first, 3), 
+        false,
+        true
     );
     REQUIRE(std::equal(bool_first, bool_last, bfirst, blast, comparator));
 
-    bit::reverse(
-        bfirst, 
-        std::next(bfirst, digits)
+    bit::replace(
+        std::next(bfirst, digits - 3), 
+        std::next(bfirst, digits + 3), 
+        bit::bit0,
+        bit::bit1
     );
-    std::reverse(
-        bool_first, 
-        std::next(bool_first, digits)
-    );
-    REQUIRE(std::equal(bool_first, bool_last, bfirst, blast, comparator));
-
-    bit::reverse(
-        std::next(bfirst, 2), 
-        std::next(bfirst, digits)
-    );
-    std::reverse(
-        std::next(bool_first, 2), 
-        std::next(bool_first, digits)
+    std::replace(
+        std::next(bool_first, digits - 3), 
+        std::next(bool_first, digits + 3), 
+        false,
+        true
     );
     REQUIRE(std::equal(bool_first, bool_last, bfirst, blast, comparator));
 
-    bit::reverse(
-        std::next(bfirst, 2), 
-        std::next(bfirst, digits - 3)
+    bit::replace(
+        std::next(bfirst, 3), 
+        std::next(bfirst, digits - 4), 
+        bit::bit0,
+        bit::bit1
     );
-    std::reverse(
-        std::next(bool_first, 2), 
-        std::next(bool_first, digits - 3)
+    std::replace(
+        std::next(bool_first, 3), 
+        std::next(bool_first, digits - 4), 
+        false,
+        true
     );
     REQUIRE(std::equal(bool_first, bool_last, bfirst, blast, comparator));
 }
 
-TEMPLATE_PRODUCT_TEST_CASE("Reverse: STD_multiple_word_main", 
+TEMPLATE_PRODUCT_TEST_CASE("Replace: multi_word", 
                            "[template][product]", 
-                           (std::vector, std::list), 
+                           (std::vector, std::list, std::forward_list), 
                            (unsigned char, unsigned short, 
                             unsigned int, unsigned long)) {
 
     using container_type = TestType;
     using num_type = typename container_type::value_type;
-    auto container_size = 10;
+    auto container_size = 8;
     auto digits = bit::binary_digits<num_type>::value;
     container_type bitcont = make_random_container<container_type>
                                      (container_size); 
@@ -101,54 +104,52 @@ TEMPLATE_PRODUCT_TEST_CASE("Reverse: STD_multiple_word_main",
     auto bool_first = std::begin(boolcont);
     auto bool_last = std::end(boolcont);
 
-    bit::reverse(
-        bfirst,
-        blast
+    bit::replace(
+        bfirst, 
+        std::next(bfirst, 3*digits), 
+        bit::bit0,
+        bit::bit1
     );
-    std::reverse(
-        bool_first,
-        bool_last
-    );
-    REQUIRE(std::equal(bool_first, bool_last, bfirst, blast, comparator));
-
-    // f.pos() <= nf.pos() and first is aligned
-    bit::reverse(
-        std::next(bfirst, 0),
-        std::next(bfirst, (container_size-4)*digits-3)
-    );
-    std::reverse(
-        std::next(bool_first, 0), 
-        std::next(bool_first, (container_size-4)*digits-3)
-     );
-    REQUIRE(std::equal(bool_first, bool_last, bfirst, blast, comparator));
-
-    // f.pos() == nf.pos() and last is aligned
-    bit::reverse(
-        std::next(bfirst, 6),
-        std::next(bfirst, (container_size-4)*digits + 6)
-    );
-    std::reverse(
-        std::next(bool_first, 6), 
-        std::next(bool_first, (container_size-4)*digits + 6)
+    std::replace(
+        bool_first, 
+        std::next(bool_first, 3*digits), 
+        false,
+        true
     );
     REQUIRE(std::equal(bool_first, bool_last, bfirst, blast, comparator));
 
-    // f.pos() <= nf.pos() and first is aligned
-    bit::reverse(
-        std::next(bfirst, 0),
-        std::next(bfirst, (container_size-4)*digits-3)
+    bit::replace(
+        std::next(bfirst, digits + 3), 
+        std::next(bfirst, 4*digits - 4), 
+        bit::bit1,
+        bit::bit0
     );
-    std::reverse(
-        std::next(bool_first, 0), 
-        std::next(bool_first, (container_size-4)*digits-3)
-     );
+    std::replace(
+        std::next(bool_first, digits + 3), 
+        std::next(bool_first, 4*digits - 4), 
+        true,
+        false
+    );
+    REQUIRE(std::equal(bool_first, bool_last, bfirst, blast, comparator));
+
+    bit::replace(
+        std::next(bfirst, digits - 1), 
+        std::next(bfirst, container_size*digits - 1), 
+        bit::bit0,
+        bit::bit1
+    );
+    std::replace(
+        std::next(bool_first, digits - 1), 
+        std::next(bool_first, container_size*digits - 1), 
+        false,
+        true
+    );
     REQUIRE(std::equal(bool_first, bool_last, bfirst, blast, comparator));
 }
-
 // -------------------------------------------------------------------------- //
 
 
 
 // ========================================================================== //
-#endif // _REVERSE_TESTS_HPP_INCLUDED
+#endif // _REPLACE_TESTS_HPP_INCLUDED
 // ========================================================================== //
