+++ src/tests/search_shift_or.cpp
@@ -1,14 +1,14 @@
 // ============================== SEARCH TESTS ============================== //
 // Project: The Experimental Bit Algorithms Library
-// Name: search.hpp
+// Name: search_shift_or.hpp
 // Description: tests for search algorithm bit iterator overloads 
 // Creator: Vincent Reverdy
 // Contributor(s): Vincent Reverdy [2019]
 //                 Bryce Kille
 // License: BSD 3-Clause License
 // ========================================================================== //
-#ifndef _SEARCH_TESTS_HPP_INCLUDED
-#define _SEARCH_TESTS_HPP_INCLUDED
+#ifndef _SEARCH_SHIFT_OR_TESTS_HPP_INCLUDED
+#define _SEARCH_SHIFT_OR_TESTS_HPP_INCLUDED
 // ========================================================================== //
 
 
@@ -16,8 +16,9 @@
 // ============================== PREAMBLE ================================== //
 // C++ standard library
 // Project sources
-#include "test_root.cc"
+#include "catch2.hpp"
 #include "bit.hpp"
+#include "test_utils.hpp"
 // Third-party libraries
 // Miscellaneous
 // ========================================================================== //
@@ -25,10 +26,11 @@
 
 
 // ----------------------------- SEARCH Tests ------------------------------- //
-TEMPLATE_PRODUCT_TEST_CASE("search: single_word pattern or smaller", 
+TEMPLATE_PRODUCT_TEST_CASE("search_shift_or: single_word pattern or smaller", 
                            "[template][product]", 
                            (std::vector), 
-                           (unsigned char)) {
+                           (unsigned char, unsigned short, 
+                            unsigned int, unsigned long)) {
 
     using container_type = TestType;
     using num_type = typename container_type::value_type;
@@ -42,23 +44,22 @@ TEMPLATE_PRODUCT_TEST_CASE("search: single_word pattern or smaller",
     auto bool_first_txt = std::begin(bool_txt);
     auto bool_last_txt = std::end(bool_txt);
 
-    // k == digits
-    auto bret = bit::search(
+    auto bret = bit::search_shift_or(
             bfirst_txt, 
             blast_txt, 
-            std::next(bfirst_txt, 4*digits),
-            std::next(bfirst_txt, 5*digits)
+            std::next(bfirst_txt, 4*digits+2),
+            std::next(bfirst_txt, 5*digits-1)
     );
     auto bool_ret = std::search(
             bool_first_txt, 
             bool_last_txt, 
-            std::next(bool_first_txt, 4*digits),
-            std::next(bool_first_txt, 5*digits)
+            std::next(bool_first_txt, 4*digits+2),
+            std::next(bool_first_txt, 5*digits-1)
     );
     REQUIRE(std::equal(bool_first_txt, bool_last_txt, bfirst_txt, blast_txt, comparator));
     REQUIRE(std::distance(bfirst_txt, bret) == std::distance(bool_first_txt, bool_ret));
 
-    bret = bit::search(
+    bret = bit::search_shift_or(
         bfirst_txt, 
         blast_txt, 
         std::next(bfirst_txt, 4*digits + 1),
@@ -73,8 +74,7 @@ TEMPLATE_PRODUCT_TEST_CASE("search: single_word pattern or smaller",
     REQUIRE(std::equal(bool_first_txt, bool_last_txt, bfirst_txt, blast_txt, comparator));
     REQUIRE(std::distance(bfirst_txt, bret) == std::distance(bool_first_txt, bool_ret));
 
-    // k < digits
-    bret = bit::search(
+    bret = bit::search_shift_or(
         bfirst_txt, 
         blast_txt, 
         std::next(bfirst_txt, 4*digits),
@@ -89,7 +89,7 @@ TEMPLATE_PRODUCT_TEST_CASE("search: single_word pattern or smaller",
     REQUIRE(std::equal(bool_first_txt, bool_last_txt, bfirst_txt, blast_txt, comparator));
     REQUIRE(std::distance(bfirst_txt, bret) == std::distance(bool_first_txt, bool_ret));
 
-    bret = bit::search(
+    bret = bit::search_shift_or(
         bfirst_txt, 
         blast_txt, 
         std::next(bfirst_txt, 4*digits + 3),
@@ -104,7 +104,7 @@ TEMPLATE_PRODUCT_TEST_CASE("search: single_word pattern or smaller",
     REQUIRE(std::equal(bool_first_txt, bool_last_txt, bfirst_txt, blast_txt, comparator));
     REQUIRE(std::distance(bfirst_txt, bret) == std::distance(bool_first_txt, bool_ret));
 
-    bret = bit::search(
+    bret = bit::search_shift_or(
         bfirst_txt, 
         blast_txt, 
         std::next(bfirst_txt, 4*digits + 3),
@@ -120,10 +120,11 @@ TEMPLATE_PRODUCT_TEST_CASE("search: single_word pattern or smaller",
     REQUIRE(std::distance(bfirst_txt, bret) == std::distance(bool_first_txt, bool_ret));
 }
 
-TEMPLATE_PRODUCT_TEST_CASE("search: multiple_word pattern | text aligned", 
+TEMPLATE_PRODUCT_TEST_CASE("search_shift_or: multiple_word pattern | text aligned", 
                            "[template][product]", 
                            (std::vector), 
-                           (unsigned char)) {
+                           (unsigned char, unsigned short, 
+                            unsigned int, unsigned long, unsigned long long)) {
 
     using container_type = TestType;
     using num_type = typename container_type::value_type;
@@ -138,7 +139,7 @@ TEMPLATE_PRODUCT_TEST_CASE("search: multiple_word pattern | text aligned",
     auto bool_last_txt = std::end(bool_txt);
 
     // Aligned
-    auto bret = bit::search(
+    auto bret = bit::search_shift_or(
             bfirst_txt, 
             blast_txt, 
             std::next(bfirst_txt, 3*digits),
@@ -154,7 +155,7 @@ TEMPLATE_PRODUCT_TEST_CASE("search: multiple_word pattern | text aligned",
     REQUIRE(std::distance(bfirst_txt, bret) == std::distance(bool_first_txt, bool_ret));
 
     //Text aligned | Pattern unaligned
-    bret = bit::search(
+    bret = bit::search_shift_or(
         bfirst_txt, 
         blast_txt, 
         std::next(bfirst_txt, 3*digits),
@@ -169,7 +170,7 @@ TEMPLATE_PRODUCT_TEST_CASE("search: multiple_word pattern | text aligned",
     REQUIRE(std::equal(bool_first_txt, bool_last_txt, bfirst_txt, blast_txt, comparator));
     REQUIRE(std::distance(bfirst_txt, bret) == std::distance(bool_first_txt, bool_ret));
 
-    bret = bit::search(
+    bret = bit::search_shift_or(
         bfirst_txt, 
         blast_txt, 
         std::next(bfirst_txt, 3*digits + 3),
@@ -184,7 +185,7 @@ TEMPLATE_PRODUCT_TEST_CASE("search: multiple_word pattern | text aligned",
     REQUIRE(std::equal(bool_first_txt, bool_last_txt, bfirst_txt, blast_txt, comparator));
     REQUIRE(std::distance(bfirst_txt, bret) == std::distance(bool_first_txt, bool_ret));
 
-    bret = bit::search(
+    bret = bit::search_shift_or(
         bfirst_txt, 
         blast_txt, 
         std::next(bfirst_txt, 3*digits + 3),
@@ -200,10 +201,11 @@ TEMPLATE_PRODUCT_TEST_CASE("search: multiple_word pattern | text aligned",
     REQUIRE(std::distance(bfirst_txt, bret) == std::distance(bool_first_txt, bool_ret));
 }
 
-TEMPLATE_PRODUCT_TEST_CASE("search: multiple_word pattern | pattern aligned", 
+TEMPLATE_PRODUCT_TEST_CASE("search_shift_or: multiple_word pattern | pattern aligned", 
                            "[template][product]", 
                            (std::vector), 
-                           (unsigned char)) {
+                           (unsigned char, unsigned short, 
+                            unsigned int)) {
 
     using container_type = TestType;
     using num_type = typename container_type::value_type;
@@ -218,7 +220,7 @@ TEMPLATE_PRODUCT_TEST_CASE("search: multiple_word pattern | pattern aligned",
     auto bool_last_txt = std::end(bool_txt);
 
     //Text unaligned | Pattern aligned
-    auto bret = bit::search(
+    auto bret = bit::search_shift_or(
         bfirst_txt, 
         std::next(bfirst_txt, container_size*digits - 3), 
         std::next(bfirst_txt, 3*digits),
@@ -233,7 +235,7 @@ TEMPLATE_PRODUCT_TEST_CASE("search: multiple_word pattern | pattern aligned",
     REQUIRE(std::equal(bool_first_txt, bool_last_txt, bfirst_txt, blast_txt, comparator));
     REQUIRE(std::distance(bfirst_txt, bret) == std::distance(bool_first_txt, bool_ret));
 
-    bret = bit::search(
+    bret = bit::search_shift_or(
         std::next(bfirst_txt, 3), 
         blast_txt, 
         std::next(bfirst_txt, 3*digits),
@@ -248,7 +250,7 @@ TEMPLATE_PRODUCT_TEST_CASE("search: multiple_word pattern | pattern aligned",
     REQUIRE(std::equal(bool_first_txt, bool_last_txt, bfirst_txt, blast_txt, comparator));
     REQUIRE(std::distance(bfirst_txt, bret) == std::distance(bool_first_txt, bool_ret));
 
-    bret = bit::search(
+    bret = bit::search_shift_or(
         std::next(bfirst_txt, 3), 
         std::next(bfirst_txt, container_size*digits - 3), 
         std::next(bfirst_txt, 3*digits),
@@ -264,10 +266,11 @@ TEMPLATE_PRODUCT_TEST_CASE("search: multiple_word pattern | pattern aligned",
     REQUIRE(std::distance(bfirst_txt, bret) == std::distance(bool_first_txt, bool_ret));
 }
 
-TEMPLATE_PRODUCT_TEST_CASE("search: multiple_word pattern | unaligned", 
+TEMPLATE_PRODUCT_TEST_CASE("search_shift_or: multiple_word pattern | unaligned", 
                            "[template][product]", 
                            (std::vector), 
-                           (unsigned char)) {
+                           (unsigned char, unsigned short, 
+                            unsigned int)) {
 
     using container_type = TestType;
     using num_type = typename container_type::value_type;
@@ -281,7 +284,7 @@ TEMPLATE_PRODUCT_TEST_CASE("search: multiple_word pattern | unaligned",
     auto bool_first_txt = std::begin(bool_txt);
     auto bool_last_txt = std::end(bool_txt);
 
-    auto bret = bit::search(
+    auto bret = bit::search_shift_or(
         bfirst_txt, 
         std::next(bfirst_txt, container_size*digits - 3), 
         std::next(bfirst_txt, 3*digits + 2),
@@ -296,7 +299,7 @@ TEMPLATE_PRODUCT_TEST_CASE("search: multiple_word pattern | unaligned",
     REQUIRE(std::equal(bool_first_txt, bool_last_txt, bfirst_txt, blast_txt, comparator));
     REQUIRE(std::distance(bfirst_txt, bret) == std::distance(bool_first_txt, bool_ret));
 
-    bret = bit::search(
+    bret = bit::search_shift_or(
         std::next(bfirst_txt, 3), 
         blast_txt, 
         std::next(bfirst_txt, 3*digits + 2),
@@ -311,7 +314,7 @@ TEMPLATE_PRODUCT_TEST_CASE("search: multiple_word pattern | unaligned",
     REQUIRE(std::equal(bool_first_txt, bool_last_txt, bfirst_txt, blast_txt, comparator));
     REQUIRE(std::distance(bfirst_txt, bret) == std::distance(bool_first_txt, bool_ret));
 
-    bret = bit::search(
+    bret = bit::search_shift_or(
         std::next(bfirst_txt, 3), 
         std::next(bfirst_txt, container_size*digits - 3), 
         std::next(bfirst_txt, 3*digits + 4),
@@ -328,10 +331,11 @@ TEMPLATE_PRODUCT_TEST_CASE("search: multiple_word pattern | unaligned",
     REQUIRE(std::distance(bfirst_txt, bret) == std::distance(bool_first_txt, bool_ret));
 }
 
-TEMPLATE_PRODUCT_TEST_CASE("search: many_word pattern | unaligned", 
+TEMPLATE_PRODUCT_TEST_CASE("search_shift_or: many_word pattern | unaligned", 
                            "[template][product]", 
                            (std::vector), 
-                           (unsigned char)) {
+                           (unsigned char, unsigned short, 
+                            unsigned int)) {
 
     using container_type = TestType;
     using num_type = typename container_type::value_type;
@@ -345,7 +349,7 @@ TEMPLATE_PRODUCT_TEST_CASE("search: many_word pattern | unaligned",
     auto bool_first_txt = std::begin(bool_txt);
     auto bool_last_txt = std::end(bool_txt);
 
-    auto bret = bit::search(
+    auto bret = bit::search_shift_or(
         bfirst_txt, 
         std::next(bfirst_txt, container_size*digits - 3), 
         std::next(bfirst_txt, 30*digits + 2),
@@ -360,7 +364,7 @@ TEMPLATE_PRODUCT_TEST_CASE("search: many_word pattern | unaligned",
     REQUIRE(std::equal(bool_first_txt, bool_last_txt, bfirst_txt, blast_txt, comparator));
     REQUIRE(std::distance(bfirst_txt, bret) == std::distance(bool_first_txt, bool_ret));
 
-    bret = bit::search(
+    bret = bit::search_shift_or(
         std::next(bfirst_txt, 3), 
         blast_txt, 
         std::next(bfirst_txt, 30*digits + 2),
@@ -375,7 +379,7 @@ TEMPLATE_PRODUCT_TEST_CASE("search: many_word pattern | unaligned",
     REQUIRE(std::equal(bool_first_txt, bool_last_txt, bfirst_txt, blast_txt, comparator));
     REQUIRE(std::distance(bfirst_txt, bret) == std::distance(bool_first_txt, bool_ret));
 
-    bret = bit::search(
+    bret = bit::search_shift_or(
         std::next(bfirst_txt, 3), 
         std::next(bfirst_txt, container_size*digits - 3), 
         std::next(bfirst_txt, 30*digits + 4),
@@ -396,5 +400,5 @@ TEMPLATE_PRODUCT_TEST_CASE("search: many_word pattern | unaligned",
 
 
 // ========================================================================== //
-#endif // _SEARCH_TESTS_HPP_INCLUDED
+#endif // _SEARCH_SHIFT_OR_TESTS_HPP_INCLUDED
 // ========================================================================== //
