+++ src/tests/swap_ranges.cpp
@@ -15,8 +15,9 @@
 // ============================== PREAMBLE ================================== //
 // C++ standard library
 // Project sources
-#include "test_root.cc"
-#include "../../ext/bit/bit.hpp"
+#include "catch2.hpp"
+#include "bit.hpp"
+#include "test_utils.hpp"
 // Third-party libraries
 // Miscellaneous
 // ========================================================================== //
