+++ src/tests/test_root.cpp
@@ -18,35 +18,35 @@
 #include <list>
 #include <forward_list>
 // Project sources
-#include "utils/test_utils.hpp"
-#include "bit_algorithm_details.hpp"
-#include "bit_details.hpp"
-#include "input_container.hpp"
-#include "input_iterator.hpp"
-// Test sources
-#include "all_of.hpp"
-#include "any_of.hpp"
-#include "copy.hpp"
-#include "count.hpp"
-#include "equal.hpp"
-#include "find.hpp"
-#include "fill.hpp"
-#include "for_each.hpp"
-#include "for_each_n.hpp"
-#include "is_heap.hpp"
-#include "is_sorted.hpp"
-#include "max_element.hpp"
-#include "mismatch.hpp"
-#include "none_of.hpp"
-#include "padded_read.hpp"
-#include "reverse.hpp"
-//#include "replace.hpp"
-#include "rotate.hpp"
-#include "sample.hpp"
-#include "search.hpp"
-#include "search_shift_or.hpp"
-#include "shift.hpp"
-#include "swap_ranges.hpp"
-#include "transform.hpp"
+//#include "utils/test_utils.hpp"
+//#include "bit_algorithm_details.hpp"
+//#include "bit_details.hpp"
+//#include "input_container.hpp"
+//#include "input_iterator.hpp"
+//// Test sources
+//#include "all_of.hpp"
+//#include "any_of.hpp"
+//#include "copy.hpp"
+//#include "count.hpp"
+//#include "equal.hpp"
+//#include "find.hpp"
+//#include "fill.hpp"
+//#include "for_each.hpp"
+//#include "for_each_n.hpp"
+//#include "is_heap.hpp"
+//#include "is_sorted.hpp"
+//#include "max_element.hpp"
+//#include "mismatch.hpp"
+//#include "none_of.hpp"
+//#include "padded_read.hpp"
+//#include "reverse.hpp"
+////#include "replace.hpp"
+//#include "rotate.hpp"
+//#include "sample.hpp"
+//#include "search.hpp"
+//#include "search_shift_or.hpp"
+//#include "shift.hpp"
+//#include "swap_ranges.hpp"
+//#include "transform.hpp"
 // Third party libraries
 // ========================================================================== //
